+++ provider/httpserver/auth/token.go
@@ -1,8 +1,9 @@
 package auth
 
 import (
+	"crypto/subtle"
+
 	"github.com/gin-gonic/gin"
-	"slices"
 )
 
 const (
@@ -21,30 +22,42 @@ type authTokenList struct {
 
 // NewAuthToken create simple auth token provider
 // checks a predefined header for a specific token
+// Security Note: while this approach may be somewhat fine for backend, machine-to-machine authentication in a highly
+// secured and controlled environment, it is inherently as insecure as it can get. DO NOT use it as a means of authentication on web
+// environments!!
 func NewAuthToken(headerName string, key string) Provider {
 	return &authToken{
 		headerName: headerName,
 		key:        key,
 	}
 }
+
+// CanAccess returns true if request is valid
+// Note: this method supports empty keys as a means to disable authentication
 func (a *authToken) CanAccess(c *gin.Context) bool {
-	if len(a.key) > 0 {
-		return c.Request.Header.Get(a.headerName) == a.key
-	}
-	return true
+	return subtle.ConstantTimeCompare([]byte(c.Request.Header.Get(a.headerName)), []byte(a.key)) == 1
 }
 
 // NewAuthTokenList create simple auth token provider
 // checks if a predefined header has a specific token from a token list
+// Security Note: while this approach may be somewhat fine for backend, machine-to-machine authentication in a highly
+// secured and controlled environment, it is inherently as insecure as it can get. DO NOT use it as a means of authentication on web
+// environments!!
 func NewAuthTokenList(headerName string, keyList []string) Provider {
 	return &authTokenList{
 		headerName: headerName,
 		keyList:    keyList,
 	}
 }
+
+// CanAccess returns true if request is valid
+// Note: this method supports empty keys as a means to disable authentication
 func (a *authTokenList) CanAccess(c *gin.Context) bool {
-	if len(a.keyList) > 0 {
-		return slices.Contains(a.keyList, c.Request.Header.Get(a.headerName))
+	key := c.Request.Header.Get(a.headerName)
+	for _, existingKey := range a.keyList {
+		if subtle.ConstantTimeCompare([]byte(key), []byte(existingKey)) == 1 {
+			return true
+		}
 	}
-	return true
+	return false
 }